Fix: Potential uninitialized variable in ParseEscape
I realize this is a reference implementation and it doesn't necessarily
guarantee correctness or non-crashy behavior -- but it seems to me that
if the code encounters a short param list in the escpae sequence,
undefined/weird/non-deterministic behavior may result. This commit at
least makes it so that if there is a short escape param count, everything
defaults to 0.