Bugfix for IME support incorrectly decoding
If you try to enter `konnichiha`, only the  `nn` as in the Japanese
hiragana ん (0x3093), turns into 0xB7. According to martins:

"for some codepages 0x81..0xfe bytes are DBCS lead bytes, where you then use extra bytes to decode codepoint"

is the leading theory. This is reproducible with other hiragana
like `wo`, but only when entered via IME and not via autohotkey.

As first noticed by `@Coda` in the Handmade Network Discord